--- src/component/settingsTutor/buttonSettingsTutor/ButtonCreateAssocationQuizLesson.tsx
@@ -1,31 +1,0 @@
-import { useState } from 'react';
-import Button from 'react-bootstrap/Button';
-import Modal from 'react-bootstrap/Modal';
-import CreateAssociationQuizLesson from '../createAssociationQuizLesson/CreateAssociationQuizLesson';
-
-
-const ButtonCreateAssociationQuizLesson = () => {
-    const [show, setShow] = useState(false);
-
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
-    return (
-        <>
-            <Button variant="success" onClick={handleShow} style={{ margin: "0.5em" }}>
-                Add <i className="bi bi-person-fill-add"></i>
-            </Button>
-
-
-            <Modal show={show}
-                onHide={handleClose}
-                backdrop="static"
-                keyboard={false}>
-                <Modal.Body>
-                    <CreateAssociationQuizLesson />
-                </Modal.Body>
-            </Modal>
-        </>
-    );
-}
-
-export default ButtonCreateAssociationQuizLesson;
